multi: Move mining code into mining package.
This contains the changes necessary to move the
mining code into the mining package.  The motivation
behind these changes is to decouple the block template
generation from other modules in the codebase to
simplify extending it with new features and tests.
