+++ mining/mining_test.go
@@ -1,9 +1,9 @@
 // Copyright (c) 2016 The btcsuite developers
-// Copyright (c) 2015-2019 The Decred developers
+// Copyright (c) 2015-2020 The Decred developers
 // Use of this source code is governed by an ISC
 // license that can be found in the LICENSE file.
 
-package main
+package mining
 
 import (
 	"container/heap"
