+++ mining/miningerror.go
@@ -1,8 +1,8 @@
-// Copyright (c) 2015-2016 The Decred developers
+// Copyright (c) 2015-2020 The Decred developers
 // Use of this source code is governed by an ISC
 // license that can be found in the LICENSE file.
 
-package main
+package mining
 
 import (
 	"fmt"
